+++ src/Building.cpp
@@ -1,7 +1,7 @@
 /*
   City4CFD
  
-  Copyright (c) 2021-2023, 3D Geoinformation Research Group, TU Delft
+  Copyright (c) 2021-2024, 3D Geoinformation Research Group, TU Delft
 
   This file is part of City4CFD.
 
@@ -42,27 +42,13 @@ Building::Building()
         : PolyFeature(1), _elevation(-global::largnum), _height(-global::largnum),
           _ptsPtr(std::make_shared<Point_set_3>()), _hasFailed(false) {}
 
-Building::Building(const int internalID)
-        : PolyFeature(1, internalID), _elevation(-global::largnum), _height(-global::largnum),
-          _ptsPtr(std::make_shared<Point_set_3>()), _hasFailed(false) {}
-
 Building::Building(const nlohmann::json& poly)
         : PolyFeature(poly, true, 1), _elevation(-global::largnum), _height(-global::largnum),
           _ptsPtr(std::make_shared<Point_set_3>()), _hasFailed(false) {}
         // 'true' here to check for polygon simplicity
 
-Building::Building(const nlohmann::json& poly, const int internalID)
-        : PolyFeature(poly, true, 1, internalID), _elevation(-global::largnum), _height(-global::largnum),
-          _ptsPtr(std::make_shared<Point_set_3>()), _hasFailed(false) {}
-        // 'true' here to check for polygon simplicity
-
 Building::Building(const Polygon_with_attr& poly)
         : PolyFeature(poly, true, 1), _elevation(-global::largnum), _height(-global::largnum),
-          _ptsPtr(std::make_shared<Point_set_3>()), _hasFailed(false) {}
-        // 'true' here to check for polygon simplicity
-
-Building::Building(const Polygon_with_attr& poly, const int internalID)
-        : PolyFeature(poly, true, 1, internalID), _elevation(-global::largnum), _height(-global::largnum),
           _ptsPtr(std::make_shared<Point_set_3>()), _hasFailed(false) {}
         // 'true' here to check for polygon simplicity
 
@@ -273,6 +259,10 @@ double Building::sq_max_dim() {
     dims.emplace_back(_elevation * _elevation);
 
     return *(std::max_element(dims.begin(), dims.end()));
+}
+
+PointSet3Ptr Building::get_points() const {
+    return _ptsPtr;
 }
 
 void Building::get_cityjson_info(nlohmann::json& b) const {