+++ src/PointCloud.cpp
@@ -1,7 +1,7 @@
 /*
   City4CFD
  
-  Copyright (c) 2021-2023, 3D Geoinformation Research Group, TU Delft
+  Copyright (c) 2021-2024, 3D Geoinformation Research Group, TU Delft
 
   This file is part of City4CFD.
 
@@ -110,7 +110,6 @@ void PointCloud::smooth_terrain() {
     for (auto& pt : mesh.points()) {
         _pointCloudTerrain.insert(pt);
     }
-    _pointCloudTerrain.add_property_map<bool> ("is_building_point", false);
 }
 
 void PointCloud::remove_points_in_polygon(const BuildingsPtr& features) {
@@ -155,7 +154,6 @@ void PointCloud::create_flat_terrain(const PolyFeaturesPtr& lsFeatures) {
             _pointCloudTerrain.insert(Point_3(pt.x(), pt.y(), 0.0));
         }
     }
-    _pointCloudTerrain.add_property_map<bool> ("is_building_point", false);
 }
 
 void PointCloud::set_flat_terrain() {
@@ -164,11 +162,11 @@ void PointCloud::set_flat_terrain() {
         flatPC.insert(Point_3(pt.x(), pt.y(), 0.));
     }
     _pointCloudTerrain = flatPC;
-    _pointCloudTerrain.add_property_map<bool> ("is_building_point", false);
 }
 
 void PointCloud::flatten_polygon_pts(const PolyFeaturesPtr& lsFeatures,
-                                     std::vector<EPECK::Segment_3>& constrainedEdges) {
+                                     std::vector<EPECK::Segment_3>& constrainedEdges,
+                                     std::vector<std::pair<Polygon_with_holes_2, int>>& newPolys) {
     std::cout << "\n    Flattening surfaces" << std::endl;
     std::map<int, Point_3> flattenedPts;
 
@@ -199,10 +197,12 @@ void PointCloud::flatten_polygon_pts(const PolyFeaturesPtr& lsFeatures,
         auto ita = Config::get().flattenSurfaces.find(f->get_output_layer_id());
         if (ita != Config::get().flattenSurfaces.end()) {
             // flatten points
-            if(f->flatten_polygon_inner_points(_pointCloudTerrain, flattenedPts, searchTree, pointCloudConnectivity)) {
+            bool isNextToBuilding = false;
+            if(f->flatten_polygon_inner_points(_pointCloudTerrain, flattenedPts, searchTree, pointCloudConnectivity,
+                                               constrainedEdges, newPolys, isNextToBuilding)) {
                 // add to list if constructing vertical borders
                 if (std::find(Config::get().flattenVertBorder.begin(), Config::get().flattenVertBorder.end(),
-                              f->get_output_layer_id()) != Config::get().flattenVertBorder.end()) {
+                              f->get_output_layer_id()) != Config::get().flattenVertBorder.end() && !isNextToBuilding) {
                     vertBorders.push_back(f);
                 }
             }
@@ -243,8 +243,9 @@ void PointCloud::buffer_flat_edges(const PolyFeaturesPtr& avgFeatures,
     std::vector<double> offsets{0.001};
     std::vector<Polygon_3> polyList;
     #pragma omp parallel for
-    for (auto& f: avgFeatures) {
-        auto& poly = f->get_poly().outer_boundary();
+    for (int i = 0; i < avgFeatures.size(); ++i) {
+    //for (auto& f: avgFeatures) { // MSVC doesn't like range loop with OMP
+        auto& poly = avgFeatures[i]->get_poly().outer_boundary();
         // set the frame
         boost::optional<double> margin = CGAL::compute_outer_frame_margin(poly.begin(), poly.end(), offsets.back());
         CGAL::Bbox_2 bbox = CGAL::bbox_2(poly.begin(), poly.end());
@@ -331,7 +332,6 @@ void PointCloud::read_point_clouds() {
     if (!Config::get().ground_xyz.empty()) {
         std::cout << "Reading ground points" << std::endl;
         IO::read_point_cloud(Config::get().ground_xyz, _pointCloudTerrain);
-        _pointCloudTerrain.add_property_map<bool>("is_building_point", false);
 
         std::cout << "    Points read: " << _pointCloudTerrain.size() << std::endl;
     } else {