+++ src/types.h
@@ -1,7 +1,7 @@
 /*
   City4CFD
  
-  Copyright (c) 2021-2023, 3D Geoinformation Research Group, TU Delft
+  Copyright (c) 2021-2024, 3D Geoinformation Research Group, TU Delft
 
   This file is part of City4CFD.
 
@@ -29,6 +29,14 @@
 #define CITY4CFD_TYPES_H
 
 #include "nlohmann/json.hpp"
+
+#ifndef M_PI
+#define M_PI 3.14159265358979323846264338327950288
+#endif
+
+#ifndef M_PI_2
+#define M_PI_2 1.57079632679489661923132169163975144
+#endif
 
 //-- Typedefs for smart pointers
 class Building;    class Boundary; class TopoFeature;